Day 3 Part 2 solution in Go
Pretty ugly nested loops, but worst case here is minimal and it's not leetcode.
Problem constraints make it easy to just scan ahead for all three lines and
run a loop to check for the first match, again assumption made is that first match
is good and others can be ignored.

Sample:
AoC Day 3 Part 2
2510