Day 4 Part 1 Solution
Took a bit to wrap my head around this but the solution is pretty simple.
Started out keeping everything as bytes but since the conversion to ints was
needed it just seems to save complexity to read in as strings and use the 'strings' package
instead.

Sample:
AoC Day 4 Part 1
536