Solution Day 6 Part 1 in Go
Couple ways of doing this I'm sure, thought here is to have a custom scan and
split function which creates a sliding window of four bytes moving over the
input. Little arcane to get the position, since we don't have a cursor.

Lazy trick with a map to compare uniqueness, this saves at best a couple lines
but looks nicer too so worth it here.

Sample:
AoC Day 6 Part 1
1134