Solution Day 10 Part 2 in Go
Extra work in part 1 wasn't really worth it but oh well.
Was able to get the solution a lot eariler but due to some
incorrect math in the CRT cycle the start and end of each
row was wrong causing slightly off output. Decided to waste time
being right... still have a bad trailing '.' but really not that
big of a deal.

AoC Day 10 Part 2
.