Solution for Day 11 Part 2 in Go
As usual, overthought most of the structure here and part 2 did not
really benefit from this.

Took a bit to understand what was being asked here, almost fell for
the big number pitfall but detoured into a little lesson on modular
arithmetic. Simplified some of the structure a bit.

AoC Day 11 Part 2
14106266886