Day 2 Part 2 in Nim
These are pretty close to Go in terms of size and complexity.
A lot of stdlib support here and some handy sugar to do some things
in less lines. Overall pleased, however there is probably some more
advanced things that can be done here, this is basically a
translation of the Go code.